third_party/blink/web_tests/external/wpt/gamepad/not-fully-active-expected.txt

This is a testharness.js-based test.
[FAIL] calling getGamepads() in a non-fully-active document
  assert_equals: No gamepads should be available expected 0 but got 4
Harness: the test ran to completion.
